requirements.txt
telethon
python-docx
tzdata
